+++ dart/telegram_client/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.3.80
+- fix on error tdlib
+
+## 0.3.79
+- update readme, Breackchange from Tdlib("libtdjson.so") to Tdlib(path_tdlib: "libtdjson.so")
+
 ## 0.3.78
 - update galaxeus_lib